requirements.txt
Django==5.2
google-api-python-client==2.194.0
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.2.2
google-generativeai==0.8.6
python-telegram-bot==20.7
python-dotenv==1.2.2
whitenoise==6.12.0
requests==2.33.1
Pillow==12.2.0
django-extensions==4.1
isodate==0.7.2
